Utils: remove addJSSignalMethods
This code was added in the beginning phases of the project but is never
actively used. The point seem to be to make a difference between signals
that come from gobjects and those that come from pure js-objects.
We can easily resurrect this later if we find a reason for this.

https://bugzilla.gnome.org/show_bug.cgi?id=704537